Live tile updates using UnityEngine.WSA.Tile
