+++ Assets/MarkerMetro/Example/Scripts/WSATile.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 844eafd90aaad4c4e9b5aa97e14d5825
+timeCreated: 1442967327
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
